# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ShopFlow ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: brandId
          in: query
          schema:
            type: integer
        - name: lowStock
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsPage"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/stock:
    patch:
      operationId: adjustStock
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustment"
      responses:
        "200":
          description: Stock adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  # Customers
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Suppliers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuppliersPage"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Orders
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, sku, purchasePrice, sellingPrice, wholesalePrice, retailPrice, currentStock, minStock, gstPercent, unit, status, createdAt]
      properties:
        id:              { type: integer }
        name:            { type: string }
        sku:             { type: string }
        barcode:         { type: ["string", "null"] }
        hsnCode:         { type: ["string", "null"] }
        categoryId:      { type: ["integer", "null"] }
        categoryName:    { type: ["string", "null"] }
        brandId:         { type: ["integer", "null"] }
        brandName:       { type: ["string", "null"] }
        purchasePrice:   { type: number }
        sellingPrice:    { type: number }
        wholesalePrice:  { type: number }
        retailPrice:     { type: number }
        gstPercent:      { type: number }
        unit:            { type: string }
        currentStock:    { type: integer }
        minStock:        { type: integer }
        location:        { type: ["string", "null"] }
        status:          { type: string }
        description:     { type: ["string", "null"] }
        createdAt:       { type: string }
    ProductsPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:  { type: array, items: { $ref: "#/components/schemas/Product" } }
        total: { type: integer }
        page:  { type: integer }
        limit: { type: integer }
    ProductInput:
      type: object
      required: [name, sku, purchasePrice, sellingPrice, wholesalePrice, retailPrice, unit, currentStock, minStock, gstPercent]
      properties:
        name:           { type: string, minLength: 1 }
        sku:            { type: string }
        barcode:        { type: string }
        hsnCode:        { type: string }
        categoryId:     { type: integer }
        brandId:        { type: integer }
        purchasePrice:  { type: number }
        sellingPrice:   { type: number }
        wholesalePrice: { type: number }
        retailPrice:    { type: number }
        gstPercent:     { type: number }
        unit:           { type: string }
        currentStock:   { type: integer }
        minStock:       { type: integer }
        location:       { type: string }
        description:    { type: string }
        status:         { type: string }
    ProductUpdate:
      type: object
      properties:
        name:           { type: string }
        sku:            { type: string }
        barcode:        { type: string }
        hsnCode:        { type: string }
        categoryId:     { type: integer }
        brandId:        { type: integer }
        purchasePrice:  { type: number }
        sellingPrice:   { type: number }
        wholesalePrice: { type: number }
        retailPrice:    { type: number }
        gstPercent:     { type: number }
        unit:           { type: string }
        minStock:       { type: integer }
        location:       { type: string }
        description:    { type: string }
        status:         { type: string }
    StockAdjustment:
      type: object
      required: [type, quantity, reason]
      properties:
        type:     { type: string, enum: [increase, decrease, damage, lost, return] }
        quantity: { type: integer }
        reason:   { type: string }
        notes:    { type: string }
...
    Supplier:
      type: object
      required: [id, name, phone, supplierCode, createdAt]
      properties:
        id:           { type: integer }
        supplierCode: { type: string }
        name:         { type: string }
        phone:        { type: string }
        email:        { type: ["string", "null"] }
        address:      { type: ["string", "null"] }
        gstNumber:    { type: ["string", "null"] }
        outstanding:  { type: number }
        status:       { type: string }
        notes:        { type: ["string", "null"] }
        createdAt:    { type: string }
    SuppliersPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:  { type: array, items: { $ref: "#/components/schemas/Supplier" } }
        total: { type: integer }
        page:  { type: integer }
        limit: { type: integer }
    SupplierInput:
      type: object
      required: [name, phone]
      properties:
        name:      { type: string, minLength: 1 }
        phone:     { type: string }
        email:     { type: string }
        address:   { type: string }
        gstNumber: { type: string }
        notes:     { type: string }
    SupplierUpdate:
      type: object
      properties:
        name:      { type: string }
        phone:     { type: string }
        email:     { type: string }
        address:   { type: string }
        gstNumber: { type: string }
        status:    { type: string }
        notes:     { type: string }
...