# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ShopFlow ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  sku: string;
  barcode?: string;
  hsnCode?: string;
  categoryId?: number;
  brandId?: number;
  purchasePrice: number;
  sellingPrice: number;
  wholesalePrice: number;
  retailPrice: number;
  gstPercent: number;
  unit: string;
  currentStock: number;
  minStock: number;
  location?: string;
  description?: string;
  status?: string;
}