# lib/api-zod/src/generated/types/supplierUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ShopFlow ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface SupplierUpdate {
  name?: string;
  phone?: string;
  email?: string;
  address?: string;
  gstNumber?: string;
  status?: string;
  notes?: string;
}